add check on module version and run version
I am choosing to impose the rule that the loading module version should
be >= to the software producing the data file that is being loaded. This
basically means the module version number should track with the CMake
version with optional patches to the module version for module-specific
fixes or updates.